requirements.txt
Flask==1.1.1
pytest
pylint
pandas
wikipedia
google-cloud-language